squash! Better format
Extract commonly used api from `api.rb` to `api/job.rb`
